+++ .eslintrc.js
@@ -1,6 +1,5 @@
-const importRules = require('@fuelrats/eslint-config/core/plugin-import')
-const jsExtensions = require('@fuelrats/eslint-config/util/importExtensions')
-const tsExtensions = require('@fuelrats/eslint-config/util/importExtensionsTypescript')
+const util = require('@fuelrats/eslint-config-react/util')
+const { withAliasResolver } = require('@fuelrats/eslint-config/util/import')
 
 module.exports = {
   env: {
@@ -15,29 +14,26 @@ module.exports = {
   ],
   globals: {
     $$BUILD: 'readonly',
+    fetch: 'readonly',
   },
   rules: {
-    'jsx-a11y/no-noninteractive-element-interactions': ['off'], // We intend to enable this once we refactor certain key components.
-    'jsdoc/require-jsdoc': ['off'], // we'll get to it someday...
-    'import/order': ['error', {
-      ...importRules.rules['import/order'][1],
+    'import/order': util.extendRule('import/order', {
       'newlines-between': 'always',
-    }],
-    '@next/next/link-passhref': ['off'], // This rule is broken so just ignore it for now.
+    }),
+    ...util.disable(
+      'jsx-a11y/no-noninteractive-element-interactions', // We intend to enable this once we refactor certain key components.
+      'jsdoc/require-jsdoc', // we'll get to it someday...
+      '@next/next/link-passhref', // This rule is broken so just ignore it for now.
+    ),
   },
   settings: {
     'import/ignore': [
       '.worker.js$',
     ],
-    'import/resolver': {
-      node: {
-        extensions: jsExtensions,
-      },
-      alias: {
-        map: [['~', './src'], ['typings', './typings']],
-        extensions: tsExtensions,
-      },
-    },
+    'import/resolver': withAliasResolver([
+      ['~', './src'],
+      ['typings', './typings']
+    ]),
   },
   overrides: [
     {