+++ src/components/TagsInput.js
@@ -2,9 +2,7 @@
 /* eslint-disable
     no-magic-numbers,
     no-negated-condition,
-    no-restricted-syntax,
     no-return-assign,
-    react/jsx-no-bind,
     react/no-unsafe,
     react/state-in-constructor,
     arrow-body-style,