+++ src/pages/api/qms/queue.js
@@ -1,7 +1,3 @@
-import { HttpStatus } from '@fuelrats/web-util/http'
-import axios from 'axios'
-
-import { InternalServerAPIError } from '~/util/server/errors'
 import getEnv from '~/util/server/getEnv'
 import acceptMethod from '~/util/server/middleware/acceptMethod'
 import jsonApiRoute from '~/util/server/middleware/jsonApiRoute'
@@ -10,92 +6,54 @@ import jsonApiRoute from '~/util/server/middleware/jsonApiRoute'
 
 
 // Module constants
-const cache = {
-  count: {
-    maxAge: 10000,
-    lastCheck: 0,
-    value: 0,
-  },
-  max: {
-    maxAge: 60000,
-    lastCheck: 0,
-    value: 0,
-  },
-}
 const env = getEnv()
 
+function getQueueLength (ctx) {
+  return ctx.fetch({
+    url: `${env.qms.url}/api/v1/queue/`,
+    headers: {
+      Authorization: `Bearer ${env.qms.token}`,
+    },
+    cache: {
+      key: 'qms-queue-count',
+      maxAge: 10000,
+    },
+  }).then(
+    ({ data }) => {
+      return data.filter((rescue) => {
+        return !rescue.pending && !rescue.in_progress
+      }).length
+    },
+  )
+}
 
+function getMaxQueueLength (ctx) {
+  return ctx.fetch({
+    url: `${env.qms.url}/api/v1/config/`,
+    headers: {
+      Authorization: `Bearer ${env.qms.token}`,
+    },
+    cache: {
+      key: 'qms-queue-max-count',
+      maxAge: 60000,
+    },
+  }).then(
+    ({ data }) => {
+      return data.max_active_clients
+    },
+  )
+}
 
 
 export default jsonApiRoute(
   acceptMethod.GET(),
   async (ctx) => {
-    const nowTime = Date.now()
-
-    if (nowTime - cache.count.lastCheck >= cache.count.maxAge) {
-      cache.count.lastCheck = nowTime
-
-      const { data, status, statusText } = await axios.get(
-        `${env.qms.url}/api/v1/queue/`,
-        {
-          headers: {
-            Authorization: `Bearer ${env.qms.token}`,
-          },
-        },
-      )
-
-      if (status === HttpStatus.OK) {
-        const rescueQueue = data.filter((rescue) => {
-          return !rescue.pending && !rescue.in_progress
-        })
-
-        cache.count.value = rescueQueue.length
-      } else {
-        ctx.error(new InternalServerAPIError(null, {
-          internalError: {
-            url: '/api/v1/queue/',
-            status,
-            statusText,
-          },
-        }))
-      }
-    }
-
-    if (nowTime - cache.max.lastCheck >= cache.max.maxAge) {
-      cache.max.lastCheck = nowTime
-      const { data, status, statusText } = await axios.get(
-        `${env.qms.url}/api/v1/config/`,
-        {
-          headers: {
-            Authorization: `Bearer ${env.qms.token}`,
-          },
-        },
-      )
-
-      if (status === HttpStatus.OK) {
-        cache.max.value = data.max_active_clients
-      } else {
-        ctx.error(new InternalServerAPIError(null, {
-          internalError: {
-            url: '/api/v1/config/',
-            status,
-            statusText,
-          },
-        }))
-      }
-    }
-
-
     ctx.send({
       id: 'static',
       type: 'queue-statistics',
       attributes: {
-        queueLength: cache.count.value,
-        maxClients: cache.max.value,
-      },
-      meta: {
-        queueAge: nowTime - cache.count.lastCheck,
-        maxClientAge: nowTime - cache.max.lastCheck,
+        queueLength: await getQueueLength(ctx),
+        maxClients: await getMaxQueueLength(ctx),
       },
     })
   },